add valhalla service, visit handler

+++ model/movements.go
@@ -5,12 +5,12 @@ import "time"
 type Movement struct {
 	Id        int       `json:"id"`
 	UserId    int       `json:"user_id"`
-	Latitude  float32   `json:"latitude"`
-	Longitude float32   `json:"longitude"`
+	Latitude  float64   `json:"latitude"`
+	Longitude float64   `json:"longitude"`
 	CreatedAt time.Time `json:"createdAt"`
 }
 
 type MovementCreate struct {
-	Latitude  float32 `json:"latitude,string"`
-	Longitude float32 `json:"longitude,string"`
+	Latitude  float64 `json:"latitude,string"`
+	Longitude float64 `json:"longitude,string"`
 }
